major changes in roles

+++ admin/includes/viewAllUsers.php
@@ -9,8 +9,8 @@
                                     <th>Email</th>
                                     <th>IS</th>
                                     <th>FOR</th>
-                                    <th colspan="3">Options</th>
-                                    <th>Edit</th>
+                                    <th colspan="2">Options</th>
+                                    <!-- <th>Edit</th> -->
                                     <th>Delete</th>
                                     
                                 </tr>
@@ -27,6 +27,15 @@
                                     $lName = $row['last_name'];
                                     $email = substr($row['email'],0,11);
                                     $role = $row['role'];
+                                    if($role == 1){
+                                        $role = 'Admin';
+                                    }else if($role == 2){
+                                        $role = 'Content Writer';
+                                    }else if($role == -1){
+                                        $role = 'Pending';
+                                    }else if($role == 4){
+                                        $role = 'blocked';
+                                    }
                                     $roleTitle = $row['role_title'];
                                     //Print into table                    
                                     echo "<tr>";
@@ -42,12 +51,23 @@
                                     }else{
                                         echo "<td>$roleTitle</td>";
                                     }
+                                    if($row['role'] == 1){
+                                       displayContentWriterButton($usrID,$username);
+                                       displayBlockAndDelete($usrID);
+                                   }else if($row['role'] == 2){
+                                        displayAdminButton($usrID,$username);
+                                        displayBlockAndDelete($usrID);
+                                    }else if($row['role'] == 4){
+                                        displayAdminButton($usrID,$username);
+                                        displayContentWriterButton($usrID,$username);
+                                        displayDeleteButton($usrID);
+                                    }else{
+                                        echo "<td colspan='3'><a onClick=\"javascript : return confirm('Are you Sure to Delete This User?'); \" class='form-control btn btn-danger' href='users_list.php?delete=$usrID'>Delete</a></td>";
+                                    }
                                     
-                                    echo "<td class='bg-info'><a class='btn btn-primary' href='users_list.php?adminApprove=$usrID&username=$username'>Admin</a></td>";
-                                    echo "<td class='bg-success'><a class='btn btn-success' href='users_list.php?contentW=$usrID&username=$username'>Content Writer</a></td>";
-                                    echo "<td class='bg-warning'><a class='btn btn-warning' href='users_list.php?block=$usrID'>Block</a></td>";
-                                    echo "<td class='bg-info'><a class='btn btn-info' href='users_list.php?source=editUser&id=$usrID'>Edit</a></td>";
-                                    echo "<td class='bg-danger'><a onClick=\"javascript : return confirm('Are you Sure to Delete This ?'); \" class='btn btn-danger' href='users_list.php?delete=$usrID'>Delete</a></td>";
+                                    
+                                    // echo "<td class='bg-info'><a class='btn btn-info' href='users_list.php?source=editUser&id=$usrID'>Edit</a></td>";
+                                    
                                     echo "</tr>";
                                 }
                                 //delete comment
@@ -60,3 +80,6 @@
                                 blockThatUser();
                             ?>
 </table>
+<?php 
+
+?>
